news_gathering/requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==2.2.9
elasticsearch==7.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
importlib-metadata==1.3.0
jieba3k==0.35.1
kombu==4.6.7
lxml==4.4.2
more-itertools==8.0.2
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
requests-file==1.4.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
